+++ src/actions/account.js
@@ -221,7 +221,6 @@ export const {
     getLedgerKey,
     getLedgerPublicKey,
     setupRecoveryMessage,
-    setupRecoveryMessageNewAccount,
     deleteRecoveryMethod,
     checkNearDropBalance,
     claimLinkdropToAccount,
@@ -301,7 +300,6 @@ export const {
         wallet.setupRecoveryMessage.bind(wallet),
         () => showAlert()
     ],
-    SETUP_RECOVERY_MESSAGE_NEW_ACCOUNT: wallet.setupRecoveryMessageNewAccount.bind(wallet),
     DELETE_RECOVERY_METHOD: [
         wallet.deleteRecoveryMethod.bind(wallet),
         () => showAlert()
